+++ packages/miniapp-compile-config/src/setAppConfig.js
@@ -1,7 +1,6 @@
 const { dirname, resolve } = require('path');
 const {
   platformMap,
-  getAppConfig,
 } = require('miniapp-builder-shared');
 const { existsSync } = require('fs-extra');
 
@@ -16,7 +15,7 @@ const { setEntry, setMultiplePackageEntry } = require('./setEntry');
 module.exports = (
   config,
   userConfig = {},
-  { context, target, entryPath, outputPath }
+  { context, target, entryPath, outputPath, staticConfig }
 ) => {
   const platformInfo = platformMap[target];
   const {
@@ -28,17 +27,15 @@ module.exports = (
   const { rootDir, command } = context;
   const mode = command;
 
-  const appConfig = getAppConfig(rootDir, target);
-
   // Need Copy files or dir
   const needCopyList = [];
   // Record all the sub app configs
   const subAppConfigList = [];
 
   if (subPackages) {
-    setMultiplePackageEntry(config, appConfig.routes, { rootDir, target, outputPath, subAppConfigList, needCopyList });
+    setMultiplePackageEntry(config, staticConfig.routes, { rootDir, target, outputPath, subAppConfigList, needCopyList });
   } else {
-    setEntry(config, appConfig.routes, { entryPath, rootDir, target, outputPath, needCopyList });
+    setEntry(config, staticConfig.routes, { entryPath, rootDir, target, outputPath, needCopyList });
   }
 
   // Set constantDir
@@ -71,8 +68,6 @@ module.exports = (
     loaderParams.constantDir.push(dirPatterns.from)
   );
 
-  config.cache(true).mode('production').target('node');
-
   config.plugin('CopyWebpackPlugin').tap(([copyList]) => {
     return [copyList.concat(needCopyList)];
   });
@@ -103,7 +98,7 @@ module.exports = (
     {
       type: 'complie',
       subPackages,
-      appConfig,
+      appConfig: staticConfig,
       subAppConfigList,
       outputPath,
       target,