+++ packages/rax-miniapp-runtime-webpack-plugin/src/generators/pkg.js
@@ -2,7 +2,7 @@ const { resolve, dirname, join } = require('path');
 const { readJSONSync } = require('fs-extra');
 const addFileToCompilation = require('../utils/addFileToCompilation');
 
-module.exports = function(compilation, { target, command, declaredDep, source = '' }) {
+module.exports = function(compilation, { target, declaredDep, source = '' }) {
   const pkgPath = resolve(process.cwd(), 'package.json');
   const dependencies = declaredDep || readJSONSync(pkgPath, {
     encoding: 'utf-8'
@@ -12,7 +12,6 @@ module.exports = function(compilation, { target, command, declaredDep, source = 
     content: JSON.stringify({
       dependencies
     }),
-    command,
     target,
   });
 };