+++ packages/rax-miniapp-runtime-webpack-plugin/src/utils/minifyCode.js
@@ -1,10 +1,21 @@
-const terser = require('terser');
+const { transformSync } = require('@builder/swc');
 const csso = require('csso');
 const prettyData = require('pretty-data').pd;
 const { platformMap } = require('miniapp-builder-shared');
 
 function minifyJS(source) {
-  return terser.minify(source).code;
+  return transformSync(source, {
+    jsc: {
+      minify: Object.assign({
+        compress: {
+          unused: false,
+        },
+        mangle: true,
+      }),
+      target: 'es2021',
+    },
+    minify: true,
+  }).code;
 }
 
 function minifyCSS(source) {